rafactoring fade system to be future compatable
