+++ pages/index.js
@@ -3,7 +3,7 @@ import Image from "next/image";
 import { Container, Row, Col, Fade } from "react-bootstrap";
 import styles from "../styles/Home.module.scss";
 import Layout from "../components/Layout";
-import ContentFade from "../components/ContentFade";
+import FadeSet from "../components/FadeSet";
 
 export default function Home() {
   return (
@@ -32,7 +32,7 @@ export default function Home() {
         <Row className="mt-4 mb-4">
           <Col>
             <h4>Projects</h4>
-            <ContentFade>
+            <FadeSet >
               <p>
                 some elements here lorem ipsum grada poleri a ipsum lorem a mona
                 ma titi a te
@@ -41,7 +41,7 @@ export default function Home() {
                 some other elements here lorem ipsum grada poleri a ipsum lorem a mona
                 ma titi a te
               </p>
-            </ContentFade>
+            </FadeSet>
           </Col>
           <Col>
             <h4>Resume</h4>
